DCAP attestation changes to worker pool 1. WPE registration changes wrt to attestation report verification
