+++ docker-compose.yaml
@@ -22,7 +22,6 @@ services:
       context: .
       dockerfile: ./docker/Dockerfile
       args:
-        - DISPLAY=${DISPLAY:-}
         - XAUTHORITY=~/.Xauthority
         - http_proxy
         - https_proxy
