+++ tc/sgx/trusted_worker_manager/enclave/epid_signup_helper.cpp
@@ -147,19 +147,133 @@ tcf_err_t EpidSignupHelper::verify_enclave_info(const char* enclave_info,
 }  // Epid_VerifyEnclaveInfo
 
 // XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
+VerificationStatus EpidSignupHelper::verify_attestation_report(
+    const ByteArray& attestation_data,
+    const ByteArray& hex_id,
+    ByteArray& mr_enclave, ByteArray& mr_signer,
+    ByteArray& encryption_public_key_hash,
+    ByteArray& verification_key_hash) {
 
-std::string EpidSignupHelper::get_enclave_id() {
-    return this->enclave_id;
-}
+    /// Verify attestation report
+    VerificationStatus result = VERIFICATION_SUCCESS;
+    std::string att_data_string = ByteArrayToString(attestation_data);
+    /// Parse the attestation data
+    JsonValue att_data_string_parsed(json_parse_string(
+        att_data_string.c_str()));
+    if (att_data_string_parsed.value == nullptr) {
+	tcf::error::ThrowIf<tcf::error::ValueError>(true, "Parsing attestation data failed");
+        return VERIFICATION_FAILED;
+    }
+    JSON_Object* proof_object = \
+        json_value_get_object(att_data_string_parsed);
+    if (proof_object == nullptr) {
+        tcf::error::ThrowIf<tcf::error::ValueError>(true, "Get proof data json object failed");
+        return VERIFICATION_FAILED;
+    }
 
-// XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
+    const char* s_value = nullptr;
 
-std::string EpidSignupHelper::get_enclave_encryption_key() {
-    return this->enclave_encryption_key;
-}
+    s_value = json_object_dotget_string(proof_object, "ias_report_signature");
+    if (s_value == nullptr) {
+	tcf::error::ThrowIf<tcf::error::ValueError>(true,
+			"Extracting ias_report_signature from proof_data failed");
+        return VERIFICATION_FAILED;
+    }
+    const std::string proof_signature(s_value);
 
-// XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
+    /// Parse verification report
+    s_value = json_object_dotget_string(proof_object, "verification_report");
+    if (s_value == nullptr) {
+	tcf::error::ThrowIf<tcf::error::ValueError>(true,
+			"Extracting verification_report from proof_data failed");
+        return VERIFICATION_FAILED;
+    }
+    const std::string verification_report(s_value);
 
-sgx_report_data_t EpidSignupHelper::get_report_data() {
-    return this->report_data;
+    JsonValue verification_report_parsed(
+        json_parse_string(verification_report.c_str()));
+    if (verification_report_parsed.value == nullptr) {
+	tcf::error::ThrowIf<tcf::error::ValueError>(true, "Parsing verification report failed");
+        return VERIFICATION_FAILED;
+    }
+
+    JSON_Object* verification_report_object = \
+        json_value_get_object(verification_report_parsed);
+    if (verification_report_object == nullptr) {
+	tcf::error::ThrowIf<tcf::error::ValueError>(true,
+	    "Get verification report json object failed");
+        return VERIFICATION_FAILED;
+    }
+
+    s_value = json_object_dotget_string(verification_report_object,
+        "isvEnclaveQuoteBody");
+    if (s_value == nullptr) {
+	tcf::error::ThrowIf<tcf::error::ValueError>(true, "Extracting isvEnclaveQuoteBody failed");
+        return VERIFICATION_FAILED;
+    }
+    const std::string enclave_quote_body(s_value);
+
+    s_value = json_object_dotget_string(
+        verification_report_object, "epidPseudonym");
+    if (s_value == nullptr) {
+        tcf::error::ThrowIf<tcf::error::ValueError>(true, "Extracting epidPseudonym failed");
+        return VERIFICATION_FAILED;
+    }
+    const std::string epid_pseudonym(s_value);
+
+    /* Verify verification report signature
+       Verify good quote, but group-of-date is not considered ok */
+    bool r = verify_enclave_quote_status(verification_report.c_str(),
+        verification_report.length(), 1);
+    if (r != true) {
+	tcf::error::ThrowIf<tcf::error::ValueError>(true, "Verifying enclave quote failed");
+        return VERIFICATION_FAILED;
+    }
+    const char* ias_report_cert = json_object_dotget_string(
+        proof_object, "ias_report_signing_certificate");
+
+    std::vector<char> verification_report_vec(
+        verification_report.begin(), verification_report.end());
+    verification_report_vec.push_back('\0');
+    char* verification_report_arr = &verification_report_vec[0];
+
+    std::vector<char> proof_signature_vec(proof_signature.begin(),
+        proof_signature.end());
+    proof_signature_vec.push_back('\0');
+    char* proof_signature_arr = &proof_signature_vec[0];
+
+    /// verify IAS signature
+    r = verify_ias_report_signature(ias_report_cert,
+                                    verification_report_arr,
+                                    strlen(verification_report_arr),
+                                    proof_signature_arr,
+                                    strlen(proof_signature_arr));
+    if (r != true) {
+        tcf::error::ThrowIf<tcf::error::ValueError>(true, "Verifying ias report signature failed");
+        return VERIFICATION_FAILED;
+    }
+
+    /* Extract ReportData and MR_ENCLAVE from isvEnclaveQuoteBody
+       present in Verification Report */
+    sgx_quote_t* quote_body = reinterpret_cast<sgx_quote_t*>(
+        Base64EncodedStringToByteArray(enclave_quote_body).data());
+    sgx_report_body_t* report_body = &quote_body->report_body;
+    sgx_report_data_t expected_report_data = *(&report_body->report_data);
+    sgx_measurement_t mr_enclave_from_report = *(&report_body->mr_enclave);
+    sgx_measurement_t mr_signer_from_report = *(&report_body->mr_signer);
+
+    /// Convert uint8_t array to ByteArray(vector<uint8_t>)
+    ByteArray mr_enclave_bytes(std::begin(mr_enclave_from_report.m),
+        std::end(mr_enclave_from_report.m));
+    ByteArray mr_signer_bytes(std::begin(mr_signer_from_report.m),
+        std::end(mr_signer_from_report.m));
+    mr_enclave = mr_enclave_bytes;
+    mr_signer = mr_signer_bytes;
+
+
+    encryption_public_key_hash = ByteArray(std::begin(expected_report_data.d),
+                                     std::begin(expected_report_data.d)+SGX_HASH_SIZE);
+    verification_key_hash = ByteArray(std::begin(expected_report_data.d)+SGX_HASH_SIZE,
+                                     std::end(expected_report_data.d));
+    return result;
 }
